Update image resolution and sizing
I tried to update the organizers image with those uploaded on the folder for Dr. Menna and Salomey. For Issam, I used the same image as was on the website and made the image resolution sharper. I also made Nene's image sharper and aligned the image size for organizers and mentors.

Please check the heading for organizers @danielajisafe. I updated yours, Blaise (adding B.Sc. ) and Salomey (adding  Ph.D. student).

Finally I update the Organizers and program committee to reflect 9 organizers. The previous BAI image under the 'Black-in-AI-CVPR23-Social' section was also replaced with a pexel free image as we do not have copyright over that.